okokok

+++ lab13_1.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 
-int fibonacci(int x);
+long long int fibonacci(int x);
   
 
 
@@ -20,7 +20,7 @@ int main(){
 }
 
 
-int fibonacci(int x){
+long long int fibonacci(int x){
      if(x<0) return 0;
      if(x==1) return 1;
     else return fibonacci(x-1)+fibonacci(x-2);
